# server/ui/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "OpenAPI 2.0 documentation for [Mailpit](https://github.com/axllent/mailpit).",
    "title": "Mailpit API",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/axllent/mailpit"
    },
    "license": {
      "name": "MIT license",
      "url": "https://github.com/axllent/mailpit/blob/develop/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/message/{ID}/html-check": {
      "get": {
        "description": "Returns the summary of the message HTML checker.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "other"
        ],
        "summary": "HTML check",
        "operationId": "HTMLCheckParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "HTMLCheckResponse",
            "schema": {
              "$ref": "#/definitions/HTMLCheckResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
    "/api/v1/message/{ID}/link-check": {
      "get": {
        "description": "Returns the summary of the message Link checker.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "other"
        ],
        "summary": "Link check",
        "operationId": "LinkCheckParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "default": "false",
            "x-go-name": "Follow",
            "description": "Follow redirects",
            "name": "follow",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "LinkCheckResponse",
            "schema": {
              "$ref": "#/definitions/LinkCheckResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/message/{ID}/sa-check": {
      "get": {
        "description": "Returns the SpamAssassin summary (if enabled) of the message.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "other"
        ],
        "summary": "SpamAssassin check",
        "operationId": "SpamAssassinCheckParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "SpamAssassinResponse",
            "schema": {
              "$ref": "#/definitions/SpamAssassinResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/message/{ID}/spam-check": {
      "get": {
        "description": "Runs the built-in heuristic spam filter against the message and returns\nthe score and triggered rules.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "other"
        ],
        "summary": "Spam filter check",
        "operationId": "SpamFilterCheckParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "SpamFilterResponse",
            "schema": {
              "$ref": "#/definitions/SpamFilterResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "HTMLCheckResponse": {
      "description": "Response represents the HTML check response struct",
      "type": "object",
      "properties": {
        "Platforms": {
          "description": "All platforms tested, mainly for the web UI",
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "Total": {
          "$ref": "#/definitions/HTMLCheckTotal"
        },
        "Warnings": {
          "description": "List of warnings from tests",
          "type": "array",
          "items": {
            "$ref": "#/definitions/HTMLCheckWarning"
          }
        }
      },
      "x-go-name": "Response",
      "x-go-package": "github.com/axllent/mailpit/internal/htmlcheck"
    },
    "HTMLCheckResult": {
      "description": "Result struct",
      "type": "object",
      "properties": {
        "Family": {
          "description": "Family eg: Outlook, Mozilla Thunderbird",
          "type": "string"
        },
        "Name": {
          "description": "Friendly name of result, combining family, platform \u0026 version",
          "type": "string"
        },
        "NoteNumber": {
          "description": "Note number for partially supported if applicable",
          "type": "string"
        },
        "Platform": {
          "description": "Platform eg: ios, android, windows",
          "type": "string"
        },
        "Support": {
          "description": "Support [yes, no, partial]",
          "type": "string"
        },
        "Version": {
          "description": "Family version eg: 4.7.1, 2019-10, 10.3",
          "type": "string"
        }
      },
      "x-go-name": "Result",
      "x-go-package": "github.com/axllent/mailpit/internal/htmlcheck"
    },
    "HTMLCheckScore": {
      "description": "Score struct",
      "type": "object",
      "properties": {
        "Found": {
          "description": "Number of matches in the document",
          "type": "integer",
          "format": "int64"
        },
        "Partial": {
          "description": "Total percentage partially supported",
          "type": "number",
          "format": "float"
        },
        "Supported": {
          "description": "Total percentage supported",
          "type": "number",
          "format": "float"
        },
        "Unsupported": {
          "description": "Total percentage unsupported",
          "type": "number",
          "format": "float"
        }
      },
      "x-go-name": "Score",
      "x-go-package": "github.com/axllent/mailpit/internal/htmlcheck"
    },
    "HTMLCheckTotal": {
      "description": "Total weighted result for all scores",
      "type": "object",
      "properties": {
        "Nodes": {
          "description": "Total number of HTML nodes detected in message",
          "type": "integer",
          "format": "int64"
        },
        "Partial": {
          "description": "Overall percentage partially supported",
          "type": "number",
          "format": "float"
        },
        "Supported": {
          "description": "Overall percentage supported",
          "type": "number",
          "format": "float"
        },
        "Tests": {
          "description": "Total number of tests done",
          "type": "integer",
          "format": "int64"
        },
        "Unsupported": {
          "description": "Overall percentage unsupported",
          "type": "number",
          "format": "float"
        }
      },
      "x-go-name": "Total",
      "x-go-package": "github.com/axllent/mailpit/internal/htmlcheck"
    },
    "HTMLCheckWarning": {
      "description": "Warning represents a failed test",
      "type": "object",
      "properties": {
        "Category": {
          "description": "Category [css, html]",
          "type": "string"
        },
        "Description": {
          "description": "Description",
          "type": "string"
        },
        "Keywords": {
          "description": "Keywords",
          "type": "string"
        },
        "NotesByNumber": {
          "description": "Notes based on results",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "Results": {
          "description": "Test results",
          "type": "array",
          "items": {
            "$ref": "#/definitions/HTMLCheckResult"
          }
        },
        "Score": {
          "$ref": "#/definitions/HTMLCheckScore"
        },
        "Slug": {
          "description": "Slug identifier",
          "type": "string"
        },
        "Tags": {
          "description": "Tags",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "Title": {
          "description": "Friendly title",
          "type": "string"
        },
        "URL": {
          "description": "URL to caniemail.com",
          "type": "string"
        }
      },
      "x-go-name": "Warning",
      "x-go-package": "github.com/axllent/mailpit/internal/htmlcheck"
    },
    "Link": {
      "description": "Link struct",
      "type": "object",
      "properties": {
        "Status": {
          "description": "HTTP status definition",
          "type": "string"
        },
        "StatusCode": {
          "description": "HTTP status code",
          "type": "integer",
          "format": "int64"
        },
        "URL": {
          "description": "Link URL",
          "type": "string"
        }
      },
      "x-go-package": "github.com/axllent/mailpit/internal/linkcheck"
    },
    "LinkCheckResponse": {
      "description": "Response represents the Link check response",
      "type": "object",
      "properties": {
        "Errors": {
          "description": "Total number of errors",
          "type": "integer",
          "format": "int64"
        },
        "Links": {
          "description": "Tested links",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Link"
          }
        }
      },
      "x-go-name": "Response",
      "x-go-package": "github.com/axllent/mailpit/internal/linkcheck"
    },
...
    "Rule": {
      "description": "Rule struct",
      "type": "object",
      "properties": {
        "Description": {
          "description": "SpamAssassin rule description",
          "type": "string"
        },
        "Name": {
          "description": "SpamAssassin rule name",
          "type": "string"
        },
        "Score": {
          "description": "Spam rule score",
          "type": "number",
          "format": "double"
        }
      },
      "x-go-package": "github.com/axllent/mailpit/internal/spamassassin"
    },
    "SpamAssassinResponse": {
      "description": "Result is a SpamAssassin result",
      "type": "object",
      "properties": {
        "Error": {
          "description": "If populated will return an error string",
          "type": "string"
        },
        "IsSpam": {
          "description": "Whether the message is spam or not",
          "type": "boolean"
        },
        "Rules": {
          "description": "Spam rules triggered",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Rule"
          }
        },
        "Score": {
          "description": "Total spam score based on triggered rules",
          "type": "number",
          "format": "double"
        }
      },
      "x-go-name": "Result",
      "x-go-package": "github.com/axllent/mailpit/internal/spamassassin"
    },
...
    "SpamFilterRule": {
      "description": "Rule struct",
      "type": "object",
      "properties": {
        "Builtin": {
          "description": "Whether the rule is a built-in (preset) rule",
          "type": "boolean"
        },
        "Description": {
          "description": "Rule description",
          "type": "string"
        },
        "Name": {
          "description": "Rule name / ID",
          "type": "string"
        },
        "Score": {
          "description": "Rule score (negative scores reduce the total)",
          "type": "number",
          "format": "double"
        }
      },
      "x-go-name": "Rule",
      "x-go-package": "github.com/axllent/mailpit/internal/spamfilter"
    },
    "SpamFilterResponse": {
      "description": "Result is a spam filter result",
      "type": "object",
      "properties": {
        "IsSpam": {
          "description": "Whether the message is spam or not (Score \u003e= Threshold)",
          "type": "boolean"
        },
        "Rules": {
          "description": "Spam rules triggered",
          "type": "array",
          "items": {
            "$ref": "#/definitions/SpamFilterRule"
          }
        },
        "Score": {
          "description": "Total spam score based on triggered rules",
          "type": "number",
          "format": "double"
        },
        "Threshold": {
          "description": "Score at which a message is considered spam",
          "type": "number",
          "format": "double"
        }
      },
      "x-go-name": "Result",
      "x-go-package": "github.com/axllent/mailpit/internal/spamfilter"
    }
  },
  "responses": {
...
    "ErrorResponse": {
      "description": "Server error will return with a 400 status code\nwith the error message in the body",
      "schema": {
        "type": "string"
      }
    },
...
    "NotFoundResponse": {
      "description": "Not found error will return a 404 status code",
      "schema": {
        "type": "string"
      }
    },
...
  }
}